New options for the footnotes widget: link_id_prepend and back_link_id_append
They only come into play when a footnote element has an id.
The link_id_prepend option adds some text before the id, so that
"my-footnote" becomes "fn-my-footnote" for example.

The back_link_id_append adds some text after the id of the back link
(the link from the footnote to the paragraph that refers to it).