Fix 'title_widget' when the selected node contains HTML tags
With 'Soup.leaf_text', the function returns 'None' when the node
contains more than one child.  However, there are legitimate
situations where we could want to do such a thing.

For instance, consider the following title:

    <h1><code>wc</code> in language Y</h1>

In soupault.1.8.0, this title results in the use of the default title.
With this patch, the rendered titles is

    <title>wc in language Y</title>