[Plugin API] Make String.url_encode second argument a list of excluded characters rather than included characters.
My initial idea was that URL encoding could be used as a way to store names
with reserved characters in a filesystem (for example, encode _just_ slashes
to make a string with slashes in it a valid UNIX file name).
However, in practice URL encoders in browsers and libraries always encode
percent characters so encoding isn't idempotent.

In that light, ability to exclude characters from encoding is much more useful
since it has more use cases, e.g. protecting ? and & from encoding
in GET request URLs.