Remove process_pages_first config option
The original idea for it was to allow a page to serve
as a source of data for other widgets
or to make the build fail early if there are known
slow steps.
Neither use makes any sense now:
for reusable data, one can use the usual indexing mechanism,
and the site index is now available to all pages by default.
The reasoning about known slow steps is not applicable
when pages are processed in "load-parse-process-render-save" stages
rather than one by one.

The option is removed without a deprecation warning
because I couldn't find any uses of that option in the wild.
If it breaks someone's process, the config update
is trivial anyway.