Remove support for passing data betweeb plugins in persistent_data and global_data
Now that site index data is available to all pages by default
and metadata extraction is a proper part of the dependency graph,
there is much less need for this communication channel.

The post-index hook can put arbitrary data in index entries,
so there is a communication channel if that is really needed --
one a lot less problematic than a shared mutable state.

A shared mutable state has two problems:
1. It can create poorly traceable "action at a distance".
2. It makes parallel processing of pages a lot harder
   since it requires locking the shared state,
   and may defeat the purpose of parallelism
   if plugins end up spending a lot of time
   waiting to acquire the write lock.

That functionality does not seem used
in publicly-available soupault-based websites,
so I presume it's safe to remove without affecting anyone.